+++ resources/js/components/MonitoredRequests.vue
@@ -12,12 +12,15 @@ export default {
             entries: [],
             ready: false,
             requestController: new AbortController(),
-            lastEntryIndex: '',
+            pageBefore: '',
+            nextBefore: '',
+            pageCursors: [''],
+            currentPage: 0,
             hasMoreEntries: true,
             hasNewEntries: false,
-            entriesPerRequest: 50,
+            entriesPerRequest: 25,
             loadingNewEntries: false,
-            loadingMoreEntries: false,
+            loadingPage: false,
             newEntriesTimeout: null,
             newEntriesTimer: 2500,
         };
@@ -79,7 +82,7 @@ export default {
                         '?endpoint=' +
                         encodeURIComponent(this.endpointFilter) +
                         '&before=' +
-                        this.lastEntryIndex +
+                        this.pageBefore +
                         '&take=' +
                         this.entriesPerRequest,
                     null,
@@ -88,9 +91,9 @@ export default {
                 .then((response) => {
                     if (signal.aborted) return;
 
-                    this.lastEntryIndex = response.data.entries.length
+                    this.nextBefore = response.data.entries.length
                         ? _.last(response.data.entries).sequence
-                        : this.lastEntryIndex;
+                        : '';
 
                     this.hasMoreEntries = response.data.entries.length >= this.entriesPerRequest;
 
@@ -103,7 +106,7 @@ export default {
 
                     this.ready = true;
                     this.loadingNewEntries = false;
-                    this.loadingMoreEntries = false;
+                    this.loadingPage = false;
 
                     if (this.mayRetry(error, signal)) {
                         this.checkForNewEntries();
@@ -112,6 +115,10 @@ export default {
         },
 
         checkForNewEntries() {
+            if (this.currentPage !== 0) {
+                return;
+            }
+
             const {signal} = this.requestController;
 
             if (!this.endpointFilter) {
@@ -155,19 +162,57 @@ export default {
             }, this.newEntriesTimer);
         },
 
-        loadOlderEntries() {
-            this.loadingMoreEntries = true;
+        resetPagination() {
+            this.currentPage = 0;
+            this.pageBefore = '';
+            this.nextBefore = '';
+            this.pageCursors = [''];
+            this.hasMoreEntries = true;
+            this.loadingPage = false;
+        },
+
+        fetchCurrentPage(fromPrevious) {
+            this.loadingPage = true;
 
             this.loadEntries((entries) => {
-                this.entries.push(...entries);
-                this.loadingMoreEntries = false;
+                this.entries = entries;
+                this.loadingPage = false;
+
+                if (fromPrevious) {
+                    this.hasMoreEntries = true;
+                }
+
+                if (this.currentPage === 0) {
+                    this.checkForNewEntries();
+                }
             });
         },
 
+        goToNextPage() {
+            if (this.loadingPage || !this.hasMoreEntries || this.nextBefore === '') {
+                return;
+            }
+
+            clearTimeout(this.newEntriesTimeout);
+            this.pageCursors.splice(this.currentPage + 1, this.pageCursors.length, this.nextBefore);
+            this.currentPage += 1;
+            this.pageBefore = this.nextBefore;
+            this.fetchCurrentPage(false);
+        },
+
+        goToPreviousPage() {
+            if (this.loadingPage || this.currentPage === 0) {
+                return;
+            }
+
+            this.currentPage -= 1;
+            this.pageBefore = this.pageCursors[this.currentPage];
+            this.fetchCurrentPage(true);
+        },
+
         loadNewEntries() {
-            this.hasMoreEntries = true;
             this.hasNewEntries = false;
-            this.lastEntryIndex = '';
+            this.resetPagination();
             this.loadingNewEntries = true;
 
             clearTimeout(this.newEntriesTimeout);
@@ -304,19 +349,24 @@ export default {
                         </router-link>
                     </td>
                 </tr>
-
-                <tr v-if="hasMoreEntries" key="olderEntries" class="dontanimate">
-                    <td colspan="100" class="text-center card-bg-secondary py-2">
-                        <small>
-                            <a href="#" v-on:click.prevent="loadOlderEntries" v-if="!loadingMoreEntries"
-                                >Load Older Entries</a
-                            >
-                        </small>
-
-                        <small v-if="loadingMoreEntries">Loading...</small>
-                    </td>
-                </tr>
             </transition-group>
         </table>
+
+        <div
+            v-if="ready && (entries.length > 0 || currentPage > 0)"
+            class="d-flex align-items-center justify-content-between border-top px-3 py-2"
+        >
+            <small>
+                <a href="#" v-if="currentPage > 0 && !loadingPage" v-on:click.prevent="goToPreviousPage">Anterior</a>
+                <span v-else class="text-muted">Anterior</span>
+            </small>
+
+            <small v-if="loadingPage">Loading...</small>
+
+            <small>
+                <a href="#" v-if="hasMoreEntries && !loadingPage" v-on:click.prevent="goToNextPage">Próximo</a>
+                <span v-else class="text-muted">Próximo</span>
+            </small>
+        </div>
     </div>
 </template>