+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/laravel/telescope/compare/v5.23.0...5.x)
 
+* Index lists show 25 entries per page, with previous and next, instead of appending older entries.
+
 ## [v5.23.0](https://github.com/laravel/telescope/compare/v5.22.1...v5.23.0) - 2026-08-27
 
 * Escape CSP nonce attribute value by [@iliaal](https://github.com/iliaal) in https://github.com/laravel/telescope/pull/1756